Move value to its own module
